added marketplace enum to extra file
